docs: update paths after project restructure
